Wipe Script Upgrades

+++ src/bin/wipe_findora_mainnet.sh
@@ -4,11 +4,14 @@ USERNAME=$USER
 ENV=prod
 NAMESPACE=mainnet
 LIVE_VERSION=$(curl -s https://${ENV}-${NAMESPACE}.${ENV}.findora.org:8668/version | awk -F\  '{print $2}')
+FINDORAD_IMG=findoranetwork/findorad:${LIVE_VERSION}
 export ROOT_DIR=/data/findora/${NAMESPACE}
 CONTAINER_NAME=findorad
 
 # Fix permissions from possible docker changes
-sudo chown -R ${USERNAME}:${USERNAME} ${ROOT_DIR}
+if [ -d ${ROOT_DIR} ]; then
+  sudo chown -R ${USERNAME}:${USERNAME} ${ROOT_DIR}
+fi
 
 ##########################################
 # Check if container is running and stop #
@@ -25,6 +28,8 @@ fi
 ####################
 # Wipe Local Files #
 ####################
-sudo rm -r /data/findora/${NAMESPACE}
+if [ -d /data/findora/${NAMESPACE} ]; then
+  sudo rm -r /data/findora/${NAMESPACE}
+fi
 sudo rm /usr/local/bin/fn
 rm ~/.findora.env
